streamlit_app/requirements.txt
streamlit>=1.30.0
pymongo>=4.6.0
python-dotenv>=1.0.0
google-generativeai>=0.3.2
pillow>=10.0.0
streamlit-option-menu>=0.3.6
streamlit-chat>=0.1.1
streamlit-extras>=0.3.4
